Refactored IPAM forms
